# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AutoViva AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List user projects
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived]
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Credentials ───────────────────────────────────────────────────────────
...
  /runs/{runId}/report:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get evaluation report
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /runs/{runId}/ask:
    post:
      operationId: askVivaAgent
      tags: [reports, ai]
      summary: Q&A with the Viva agent about the report
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VivaQuestionInput"
      responses:
        "200":
          description: Viva agent answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VivaAnswer"

  # ── Analytics ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ────────────────────────────────────────────────────────────────
...
    Project:
      type: object
      required: [id, userId, name, baseUrl, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        baseUrl:
          type: string
        techStack:
          type: ["string", "null"]
        loginUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, archived]
        testSuiteCount:
          type: ["integer", "null"]
        runCount:
          type: ["integer", "null"]
        avgScore:
          type: ["number", "null"]
        lastRunAt:
          type: ["string", "null"]
        lastRunScore:
          type: ["integer", "null"]
        lastRunGrade:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [name, baseUrl]
      properties:
        name:
          type: string
        description:
          type: string
        baseUrl:
          type: string
        techStack:
          type: string
        loginUrl:
          type: string
        loginPageType:
          type: string
        postLoginRedirect:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        baseUrl:
          type: string
        techStack:
          type: string
        loginUrl:
          type: string
        status:
          type: string
          enum: [active, archived]

    # ── Credentials ─────────────────────────────────────────────────────────
...
    TestRun:
      type: object
      required: [id, projectId, status, mode, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        suiteId:
          type: ["string", "null"]
        userId:
          type: string
        mode:
          type: string
          enum: [live_demo, pre_check, strict, quick_smoke]
        status:
          type: string
          enum: [pending, running, completed, failed]
        score:
          type: ["integer", "null"]
        grade:
          type: ["string", "null"]
        totalTests:
          type: ["integer", "null"]
        passed:
          type: ["integer", "null"]
        failed:
          type: ["integer", "null"]
        durationSeconds:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        suiteName:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    TestResult:
      type: object
      required: [id, runId, testCaseId, title, status]
      properties:
        id:
          type: string
        runId:
          type: string
        testCaseId:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [passed, failed, skipped, error]
        stepsExecuted:
          type: ["integer", "null"]
        expectedResult:
          type: object
        actualResult:
          type: object
        errorMessage:
          type: ["string", "null"]
        durationSeconds:
          type: ["number", "null"]
        screenshots:
          type: array
          items:
            $ref: "#/components/schemas/Screenshot"
        executedAt:
          type: ["string", "null"]
    Screenshot:
      type: object
      required: [id, url]
      properties:
        id:
          type: string
        url:
          type: string
        stepNumber:
          type: ["integer", "null"]
        label:
          type: ["string", "null"]

    # ── Reports ─────────────────────────────────────────────────────────────
    Report:
      type: object
      required: [id, runId, generatedAt]
      properties:
        id:
          type: string
        runId:
          type: string
        summary:
          type: ["string", "null"]
        aiNotes:
          type: ["string", "null"]
        suggestions:
          type: array
          items:
            type: string
        scoreBreakdown:
          type: object
        keyFindings:
          type: array
          items:
            type: string
        bugsFound:
          type: array
          items:
            $ref: "#/components/schemas/BugItem"
        featureCoverage:
          type: array
          items:
            $ref: "#/components/schemas/FeatureCoverageItem"
        run:
          $ref: "#/components/schemas/TestRun"
        results:
          type: array
          items:
            $ref: "#/components/schemas/TestResult"
        generatedAt:
          type: string
          format: date-time
    BugItem:
      type: object
      required: [issue, severity]
      properties:
        issue:
          type: string
        testCaseId:
          type: ["string", "null"]
        severity:
          type: string
          enum: [critical, major, minor]
        description:
          type: string
    FeatureCoverageItem:
      type: object
      required: [feature, status]
      properties:
        feature:
          type: string
        status:
          type: string
          enum: [passed, failed, not_tested]
    VivaQuestionInput:
      type: object
      required: [question]
      properties:
        question:
          type: string
    VivaAnswer:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
        evidenceScreenshots:
          type: array
          items:
            $ref: "#/components/schemas/Screenshot"
        referencedTestCases:
          type: array
          items:
            type: string

    # ── Analytics ───────────────────────────────────────────────────────────
...